Limit the pixel brightness to be relative to the global brightness setting.
The global brightness is very useful to limit the current with large
strips, so now the individual pixel brightness is set relative to the
global brightness (rounding up to the nearest integer on the basis that
some light is expected unless the brightness is set to 0).

Also some basic bounds checking (and adjustment) is included for the
global brightness.